Lower deployment target to iOS 7.0
Our app and SDK still target iOS 7.0 so Mantle will need to target iOS 7.0
as well. I asked the developers about this and they said iOS 7 is not supported
in Mantle 2.0 but there is not much stopping us from lowering the deployment
target. I assume Mantle is iOS 8+ because they want to build dynamic frameworks.